Prepped for demo
Tested to ensure that all layers load from the new Map Service. Also,
feature extent limiting now works for all layers.

Layer stacking has been implemented successfully as well. Layers stack
in the order in which they are addded to the map. If a layer is removed,
and then added back to the map, it is added at the top of the stack.